- all OneWire measures are now float - compatibility with Sensor base class

+++ src/OneWire.cpp
@@ -63,6 +63,6 @@ void OneWire::wireSelectThisDevice() {
     pBM->wireSelect(pId);
 }
 
-bool OneWire::getValue(double &eValue) {
+bool OneWire::getValue(float &eValue) {
     return 0;
 }
